refactor: move general helpers out of sidebar dir to hooks and modules
