+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2026-06-29T10:35:31.270Z\n"
-"PO-Revision-Date: 2026-06-29T10:35:31.270Z\n"
+"POT-Creation-Date: 2026-06-29T14:27:46.075Z\n"
+"PO-Revision-Date: 2026-06-29T14:27:46.075Z\n"
 
 msgid ""
 "Some dimensions were not added because they cannot be used in a "
@@ -682,19 +682,6 @@ msgid "Category option group sets"
 msgstr "Category option group sets"
 
 msgid ""
-"{{- dataSourceTetName}} dimensions cannot be combined with {{- "
-"layoutTetName}} dimensions already in the layout."
-msgstr ""
-"{{- dataSourceTetName}} dimensions cannot be combined with {{- "
-"layoutTetName}} dimensions already in the layout."
-
-msgid "Already used as custom value."
-msgstr "Already used as custom value."
-
-msgid "Cannot be used in a {{visType}}."
-msgstr "Cannot be used in a {{visType}}."
-
-msgid ""
 "Choose a data source above to see its dimensions, periods, and organisation "
 "units."
 msgstr ""
@@ -986,6 +973,19 @@ msgstr "Enrollment status"
 
 msgid "Registration org. unit"
 msgstr "Registration org. unit"
+
+msgid ""
+"{{- dataSourceTetName}} dimensions cannot be combined with {{- "
+"layoutTetName}} dimensions already in the layout."
+msgstr ""
+"{{- dataSourceTetName}} dimensions cannot be combined with {{- "
+"layoutTetName}} dimensions already in the layout."
+
+msgid "Already used as custom value."
+msgstr "Already used as custom value."
+
+msgid "Cannot be used in a {{visType}}."
+msgstr "Cannot be used in a {{visType}}."
 
 msgid "Columns"
 msgstr "Columns"
