+++ src/hooks/index.ts
@@ -41,6 +41,8 @@ export * from './use-stable-callback'
 export * from './use-program-ids'
 export * from './use-program-stage-ids'
 export * from './use-tet-id'
+export * from './use-cross-tet-mismatch'
+export * from './use-dimension-layout-blocked-message'
 export * from './intl'
 
 export const { useMutateMutation: useRtkMutation } = api
